Update README.md
## ** Please note**: This script assumes that tutorials dictionary is in the same Python file. Also, you'll need to ensure that the first time you run save_feedback_data, it includes a header row. You could check if the file is empty, and if it is, write data.keys() as the first row.
The tutorial dictionary contains valuable resources but needed: add sentiment analysis, link for the feedback form, and contact_info field in use, provide an interactive experience for the users.

 the feedback collection more interactive:

Instead of simply printing the feedback questions, create an interactive form within the console where users can provide their answers right away. You can use the input() function in Python for this.

The contact_info field can be used to ask for the user's contact details if they're open to providing them. This could be useful for follow-up questions or clarifications.

The sentiment analysis